+++ examples/tauri-app/CHANGELOG.md
@@ -1,5 +1,14 @@
 # prometheus-entity-management-tauri
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies [7ff2fe2]
+  - @prometheus-ags/prometheus-entity-management@3.1.0
+  - @prometheus-ags/entity-graph-core@3.1.0
+  - @prometheus-ags/entity-graph-tauri@3.1.0
+
 ## 0.1.1
 
 ### Patch Changes