+++ packages/entity-graph-core/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @prometheus-ags/entity-graph-core
 
+## 3.1.0
+
 ## 3.0.5
 
 ### Patch Changes