+++ packages/entity-graph-htmx/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prometheus-ags/entity-graph-htmx
 
+## 3.1.0
+
+### Patch Changes
+
+- @prometheus-ags/entity-graph-core@3.1.0
+- @prometheus-ags/entity-graph-sdl@3.1.0
+
 ## 3.0.5
 
 ### Patch Changes